problem 8
Also added a new input/ subdirectory, which will hold the specific
inputs necessary for some problems. In this case, it's a 1000-digit
number. Problems that need input will read on stdin, so for example to
get the solution to #8 we do

  ./euler 8 <inputs/8

Note also that we changed the solution type from int to long because the
problem 8 solution overflows. Maybe use fixed-width types?

+++ solutions/map.h
@@ -3,6 +3,6 @@
 #include <functional>
 #include <map>
 
-using Solution = std::function<int()>;
+using Solution = std::function<long()>;
 
 extern std::map<int, Solution> solutions;
